requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
httpx==0.28.1

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36
alembic==1.14.1

# AI / LLM
anthropic==0.42.0

# PDF Extraction
pdfplumber==0.11.4
pytesseract==0.3.13
Pillow==11.1.0

# Fuzzy Matching
rapidfuzz==3.11.0

# Google APIs
google-auth==2.37.0
google-api-python-client==2.159.0
google-auth-oauthlib==1.2.1

# Slack
slack-bolt==1.22.0
slack-sdk==3.34.0

# Notion
notion-client==2.2.1

# Scheduling
APScheduler==3.10.4

# Logging
structlog==24.4.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-cov==6.0.0
testcontainers[postgres]==4.9.0
responses==0.25.6

# Utilities
python-multipart==0.0.20
openpyxl==3.1.5
